requirements.txt
fastapi
uvicorn
groq
pymongo
pydantic
python-dotenv